audioflinger: Fix audio for WifiDisplay
AudioFlinger is not able to determine the correct
pid/tid for WifiDisplay and thus we do not pass checks
for CAPTURE_AUDIO_OUTPUT and RECORD_AUDIO permissions.

To fix audio for WifiDisplay, it should be safe to
always allow a trusted calling uid (AID_MEDIA which
has the same perms as AID_AUDIOSERVER).

Conflicts:
       services/audiopolicy/service/AudioPolicyInterfaceImpl.cpp

Change-Id: Ifa46d8e77a43027645cad02a04263b58e134c3ad
Signed-off-by: DennySPb <dennyspb@gmail.com>
Signed-off-by: TogoFire <italomellopereira@gmail.com>